+++ packages/jss/benchmark/tests/create-instance.js
@@ -2,7 +2,7 @@ import {create} from 'jss'
 import preset from 'jss-preset-default'
 
 // Avoid memory leak with registry.
-const options = {virtual: true}
+const options = {Renderer: null}
 
 suite('create a new instance', () => {
   benchmark('without preset', () => {